implement adjustable framerate
allow the user to disable vsync and pace frame rendering manually

+++ lib/sink/omtsink/omt_disabled.go
@@ -28,4 +28,4 @@ func (f *OmtSink) Frames() *layer.FrameForwarder {
 	return &f.frames
 }
 
-func (f *OmtSink) SetRate(rate int) {}
+func (f *OmtSink) SetRate(rate float64) {}
